refactor: Move all html pages into templates
Moved all html pages into templates to do rendering dynamically.
Added MainController to render index page.
Added MvcConfig to decide orders following precedence.
Changed <a> link path to use controllers.
