+++ src/main/java/com/codessquad/qna/controller/MainController.java
@@ -1,0 +1,13 @@
+package com.codessquad.qna.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class MainController {
+
+    @RequestMapping("/")
+    public String index() {
+        return "index";
+    }
+}
