+++ src/main/resources/templates/common/navbar.html
@@ -26,7 +26,7 @@
                         <li><a href="https://facebook.com" target="_blank">Facebook</a></li>
                     </ul>
                 </li>
-                <li><a href="list.html"><i class="glyphicon glyphicon-user"></i></a></li>
+                <li><a href="/users/list"><i class="glyphicon glyphicon-user"></i></a></li>
             </ul>
         </div>
     </div>
