+++ src/main/resources/templates/common/subnav.html
@@ -5,7 +5,7 @@
                data-toggle="dropdown"><i class="glyphicon glyphicon-home" style="color:#dd1111;"></i> Home <small><i
                     class="glyphicon glyphicon-chevron-down"></i></small></a>
             <ul class="nav dropdown-menu">
-                <li><a href="/user/profile.html"><i class="glyphicon glyphicon-user" style="color:#1111dd;"></i> Profile</a>
+                <li><a href="/users/profile"><i class="glyphicon glyphicon-user" style="color:#1111dd;"></i> Profile</a>
                 </li>
                 <li class="nav-divider"></li>
                 <li><a href="#"><i class="glyphicon glyphicon-cog" style="color:#dd1111;"></i> Settings</a></li>
@@ -20,9 +20,9 @@
         </div>
         <div class="collapse navbar-collapse" id="navbar-collapse2">
             <ul class="nav navbar-nav navbar-right">
-                <li class="active"><a href="../index.html">Posts</a></li>
-                <li><a href="/user/login.html" role="button">로그인</a></li>
-                <li><a href="/user/form.html" role="button">회원가입</a></li>
+                <li class="active"><a href="/">Posts</a></li>
+                <li><a href="/users/login" role="button">로그인</a></li>
+                <li><a href="/users/form" role="button">회원가입</a></li>
                 <li><a href="#" role="button">로그아웃</a></li>
                 <li><a href="#" role="button">개인정보수정</a></li>
             </ul>
