+++ src/main/resources/templates/index.html
@@ -1,4 +1,4 @@
-{{#partial "content}}
+{{#partial "content"}}
 <div class="container" id="main">
     <div class="col-md-12 col-sm-12 col-lg-10 col-lg-offset-1">
         <div class="panel panel-default qna-list">
